requirements.txt
# Core dependencies
python-telegram-bot==20.7
python-dotenv==1.0.0
pytz==2024.1

# Database
aiosqlite==0.19.0

# Utilities
aiofiles==23.2.1
pydantic==2.5.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.0
flake8==7.0.0
mypy==1.8.0

# Optional (for production)
redis==5.0.1
uvloop==0.19.0